Usability improvements for iplot
This commit improves the usability of `iplot`. Previously, we needed to set the `fig_kw`
dictionary in order to pass keyword arguments to the backend. That's not needed anymore,
making the experience easier and similar to the other plot functions.

Also, an attempt has been made to make `InteractivePlot` inherits from `Plot`. Ultimately,
it is better to avoid this inheritance because it would require to create multiple classes,
one for each backend, structurally very similar. Furthermore, it would be much difficult
to automatically select a backend whenever the `backend` keyword argument is not provided.
This means the original implementation, in which the `InteractivePlot` class has a
`_backend` attribute, is still the best and the one to be used.