Refactoring of plotgrid
Previously, plotgrid only fully supported `MatplotlibBackend`. Now, there is full
support to all backends. If all input plots are instances of `MB`, than a matplotlib
figure will be created, otherwise an instance of `pn.GridSpec` will be returned.

In the second mode of operation, the user passes in a dictionary mapping a matplotlib
`GridSpec` object to the plots. If any of the plot is not an instance of `MB`, then
an instance of `pn.GridSpec` will be generated, maintaing the layout provided by
the user.

Finally, added tests for this functionality.