Dropped plot implicit support for Plotly and Bokeh
There are two implicit algorithms:

1. the adaptive one, which is going to create a list of rectangles to be filled by a
   solid color.
2. the meshgrid approach, which is going to use contour lines or filled contours.

Bokeh and Plotly are unable to provide an accurate representation of the numerical
data generated by an implicit series.

While matplotlib provides an extensive contour interface (which allows to set the
contour levels to be shown), Plotly's implementation is not there yet.
Bokeh doesn't even provides a way to plot proper contours: sure, it's possible to
mimick the behaviour with an Image, but that's just an approximation showing the
actual pixels with no interpolation between the values.

Also, matplotlib provides a fill command, whereas Plotly and Bokeh does not.
Said differently, it is possible to fill rectangular areas with Heatmaps or Images,
but the adaptive algorithm provides non-uniform size rectangles, making Heatmaps
or Images unsuitable for the task.

Finally, automatic line/fill coloring has been added to MatplotlibBackend's
implicit rendering.

Also modified the test to take into account the edits of this commit.