Implemented iplot sum
Implemented capability to add together multiple instances of `InteractivePlot` and `Plot`.
`InteractivePlot` instances must use the same parameters, otherwise the process is going to
fail because `panel` uses class attributes to store the parameters, which are going to be
cleared everytime a new instance is created. If two instances has the same parameters, then
the class attributes will be the same, thus no problem is encountered. If two instances has
different parameters, when the second instance is created the class attributes of the first
gets deleted, which is going to create errors.