MatplotlibBackend and plot_piecewise refactoring
Everytime a matplotlib figure is instantiated into an interactive environment, the picture will be
shown on the screen. We could use `plt.ioff()`, but then if the plot is created on a Jupyter Notebook
without `%matplotlib widget`, the picture will not be rendered on the screen.

Moreover, when building the documentation with Sphinx, it appears that each figures might be displayed
twice (random behaviour: sometimes it happens, sometimes it does not).

Hence, with `MatplotlibBackend`, the figure must be instantiated and shown at the same time. Therefore,
`plt.figure()` can't be executed inside `MatplotlibBackend.__init__()`, rather inside
`MatplotlibBackend.show()`.

This leads to another problem: that `plot_piecewise` won't be able to work with `MatplotlibBackend`.
A refactoring of `plot_piecewise` was needed. Now, each backend exposes the `_set_piecewise_color`
method, which is used to set a single color to multiple data series belonging to a single `Piecewise`
expression.