Removed tutorials from the documentation
Sad decision that had to be taken: currently, there is no reliably way to convert Jupyter
Notebooks to .rst files. Previously, the process was done manually: the pictures were generated
in a notebook, saved into `_static` folder and loaded into the manually written tutorials.
This is a pain in the ass to keep up to date.

There is `jupyter_sphinx` that generally does a good job at including output executed into
a notebook environment. Unfortunately, if different JS plotting libraries are being use (such
as in my tutorials), then there are JS clashes and the figures (mostly Plotly) won't be
visible.

At the end, it is tremendously difficult to include tutorials that shown the capabilities of
each backend into Sphinx. Hence, to keep my mental health they had to be removed.