use_cm=False won't add a colorbar on lines
Previously, a colorbar was added to every parametric line, even if use_cm=False. As a result,
when plotting many (14 or more) symbolic lines (both 2D and 3D) the horizontal space used by
the colorbars is greater than the figure. This raises a weird error.

With this commit I don't add the colorbar to parametric lines when use_cm=False. Doing so I can
add as many lines as I want.

However, the weird error will still be raised if 14 or more parametric lines are plotted with
use_cm=True. Realistically, who is gonna do that?