Improved saving support for K3D
It is now possible to use K3DBackend also on non-Jupyter consoles. Doing this allows to save the
plot in html file (png will raise errors, it only works on Jupyter).

Also, fixed color_range: it must be a Python float, not np.float32 otherwise it will raise an
error while saving the plot in html.