Refactoring of backends - Introducing renderers
Up until now, all the rendering logic was contained on a single file,
the backend.py. The biggest disadvantage is that for each data series the
logic was splitted in two methods, one for drawing and the other for updating.
This would make bug fixing and further development much more human.error prone,
because one had to scroll several hunders lines of code to modify both methods,
also identifying the correct spot was non-trivial.

Introducing renderers: a data series is now going to be associate to a particular
renderer, which implements two methods, `draw()` and `update()`. This allows
the following:

1. separate the rendering logic into different files, sharing the same structure.
   It becomes easier to maintain and upgrade the code.
2. Allows code reusability between different series/renderers.
3. Allows code extensibility: now users can implement their renderers to plot
   different kinds of symbolic expressions, pass them to the backend and use it
   with this plotting module.