Implemented workaround for holoviz's panel bug
The following issue: https://github.com/holoviz/panel/issues/6275

was preventing panel's interactive application to be opened. This commit implements
a workaround: the __init__ method of the custom template class was copied
from panel's base template and adjusted to avoid the bug to appear.

This workaround needs to be "removed" once the aforementioned issue is resolved.