Introduced Lambdifier class
In the previous iteration of the module, a symbolic expression would have
been lambdify twice:

1. the first with the requested user module (default to numpy/scipy),
2. the second using sympy, which would only be used if the evaluation with
   the first lambda function failed.

Obviously, the larger the expression the more time it took to lambdifying
them. The `Lambdifier` class does what the name suggests: it keeps track
of the data series' expressions and color_func attribute, and it generates
the appropriate lambda function only when it is requested.

Here are the steps that take place with this new setup:

1. User instantiate the data series, which holds an instance of Lambdifier.
2. User executes the `get_data()` method.
3. The evaluation algorithms request the lambda functions to Lambdifier,
   using the user-provided modules (default to Numpy/Scipy).
4. The numerical evaluation is attempted. If everything goes all right,
   it post-process the results and finally returns them.
5. If something goes wrong, it request new lambda functions to Lambdifier,
   this time using sympy as the evaluation module.
6. Same as point 4.

The generated lambda functions are kept in an appropriate dictionary, so
whenever it is requested, first it checks if it was already generated.

The same principles are used for the ``color_func`` attribute.