Remove the adaptive module
The adaptive module was implemented as an attempt for back compatibility with the old
sympy.plotting, in view of an eventual merger of spb into sympy. That never materialized.
Considering that I never found a single occasion in which the adaptive evaluation was
better than the uniform grid evaluation, I decided to remove it alltogether.
The main advantages of this choice are:

1. easier to use plotting module, as there are 3/4 keyword arguments less to be
   worried about.
2. cleaner and simpler code base.
3. easier to implement new features (interactive ones), that only works with
   the uniform evaluation strategy.