Fixed bug with priority of rcParams
In version 4.0.0 I added theme="default", which was then going to be applied
every time a new figure was instantiated. Turns out that setting a theme with
`plt.style.context(theme)` is going to disregard user-provided settings with
`plt.rcParams`. This commit sets theme="", which would still use Matplotlib
default theme, but it will be loaded through rcParams. So any user-provided
settings to rcParams will be applied to the figure, provided that theme="".